Add a test case for ENG-511 and fix a bug exposed by the test case.

+++ src/frontend/org/voltdb/exceptions/EEException.java
@@ -52,7 +52,7 @@ public class EEException extends SerializableException {
     }
 
     @Override
-    protected short p_getSerializedSize() {
+    protected int p_getSerializedSize() {
         return 4;
     }
